requirements.txt
agno
python-dotenv
pymupdf
pandas
python-docx
